tests/wpt/tests/css/css-flexbox/balance/flex-wrap-valid.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-flexbox-2/#flex-wrap-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("flex-wrap", "nowrap");
test_valid_value("flex-wrap", "wrap");
test_valid_value("flex-wrap", "wrap-reverse");
test_valid_value("flex-wrap", "balance");
test_valid_value("flex-wrap", "wrap balance", "balance");
test_valid_value("flex-wrap", "wrap-reverse balance");
test_valid_value("flex-wrap", "balance wrap", "balance");
test_valid_value("flex-wrap", "balance wrap-reverse", "wrap-reverse balance");
</script>
</body>
